Drop autotools-based build
Our official instructions only tell how to build with Meson, our CI is
only done with Meson and I wouldn't be surprised if most developers only
used Meson as well.

In other words, the autotools build is outdated. Let's say our goodbye
and drop it in the next stable release.